requirements/requirements.txt
Django==1.6.10
South==1.0
argparse==1.2.1
django-admin-honeypot==0.3.0
django-axes==1.3.6
django-dbbackup==2.0.3
django-bootstrap-pagination==1.5.0
django-bootstrap-toolkit==2.15.0
django-bootstrap3==4.11.0
django-braces==1.4.0
django-cache-utils==0.7.2
django-chroniker==0.6.7
django-cities-light==2.4.2
django-clear-cache==0.3
django-countries==2.1.2
django-country-blocker==1.1.12
django-country-dialcode==0.5.1
django-currencies==0.3.2
django-database-size==0.3.0
django-datetime-widget==0.9
django-google-charts==0.1.1
django-iban==0.2.8
django-import-export==0.2.3
django-jsonfield==0.9.13
django-qsstats-magic==0.7.2
django-queryset-csv==0.2.9
django-mass-edit==2.2
django-mathfilters==0.3.0
django-model-report==0.2.1
django-request==1.0.1
django-secure==1.0
django-simple-import==1.16
django-simple-invoice>=0.3.4
django-system-monitor==0.5
django-tables2==0.15.0
django-templated-email==0.4.9
dnspython>=1.12.0
easy-thumbnails>=2.2
GeoIP>=1.3.2
netaddr>=0.7.12
netifaces>=0.10.4
psutil>=2.1.1
psycopg2>=2.5.4
python-memcached>=1.53
raven>=3
requests>=2.4.0
pytz>=2014.4
suds>=0.4
switch2bill-common==2.8.5
vatnumber>=1.2
wsgiref==0.1.2
yawd-admin==0.7.0